% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/async.R
\name{async}
\alias{async}
\title{Construct an async function}
\usage{
async(fn)
}
\arguments{
\item{fn}{An anonymous function within which \code{await()} calls are
allowed.}
}
\value{
A function that returns a \code{\link[promises:promise]{promises::promise()}}.
}
\description{
Construct an async function
}
