% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/async.R
\name{async_generator}
\alias{async_generator}
\title{Construct an async generator}
\usage{
async_generator(fn)
}
\arguments{
\item{fn}{An anonymous function describing an async generator
within which \code{await()} calls are allowed.}
}
\value{
A generator factory. Generators constructed with this
factory always return \code{\link[promises:promise]{promises::promise()}}.
}
\description{
Construct an async generator
}
