% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/async.R
\name{async_sleep}
\alias{async_sleep}
\title{Sleep asynchronously}
\usage{
async_sleep(seconds)
}
\arguments{
\item{seconds}{The number of second to sleep.}
}
\value{
A chainable promise.
}
\description{
Sleep asynchronously
}
