% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step-reduce.R
\name{box_reduced}
\alias{box_reduced}
\alias{ensure_reduced}
\title{Create a boxed value for early termination of reduction}
\usage{
box_reduced(x)

ensure_reduced(x)
}
\arguments{
\item{x}{A value to box.}
}
\description{
A \link[rlang:box]{boxed} value of class \code{reduced} signals early
termination to \code{\link[=reduce_steps]{reduce_steps()}}. The boxed value is unboxed and
returned right away to the caller of \code{reduce_steps()}.
\itemize{
\item \code{box_reduced()} always boxes its input in a new box.
\item \code{ensure_reduced()} first checks if its input is a box of class
\code{reduced}. If it isn't, it boxes the input. Otherwise the input
is returned as is. This is useful to avoid double-boxing a value.
}
}
\examples{
box <- box_reduced(letters)

# Use `is_box(x, "reduced")` to check for a boxed value of type
# "reduced"
rlang::is_box(box, "reduced")
}
