% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coro-package.R
\docType{package}
\name{coro-package}
\alias{coro}
\alias{coro-package}
\title{coro: 'Coroutines' for R}
\description{
Provides 'coroutines' for R, a family of functions that can be suspended and resumed later on. This includes 'async' functions (which await) and generators (which yield). 'Async' functions are based on the concurrency framework of the 'promises' package. Generators are based on a dependency free iteration protocol defined in 'coro' and are compatible with iterators from the 'reticulate' package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-lib/coro}
  \item \url{https://coro.r-lib.org/}
  \item Report bugs at \url{https://github.com/r-lib/coro/issues}
}

}
\author{
\strong{Maintainer}: Lionel Henry \email{lionel@posit.co}

Authors:
\itemize{
  \item Lionel Henry \email{lionel@posit.co}
}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
