% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine-loop.R
\name{coro_advance}
\alias{coro_advance}
\title{Operators for coroutine state machines}
\usage{
coro_advance(elt_sym, iterator, env = caller_env())
}
\description{
Operators for coroutine state machines
}
\keyword{internal}
