% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{coro_debug}
\alias{coro_debug}
\title{Debug a generator or async function}
\usage{
coro_debug(fn, value = TRUE)
}
\arguments{
\item{fn}{A generator factory or an async function.}

\item{value}{Whether to debug the function.}
}
\description{
\itemize{
\item Call \code{coro_debug()} on a \code{\link[=generator]{generator()}}, \code{\link[=async]{async()}}, or
\code{\link[=async_generator]{async_generator()}} function to enable step-debugging.
\item Alternatively, set \code{options(coro_debug = TRUE)} for
step-debugging through all functions created with coro.
}
}
