% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine-switch.R
\name{coro_goto}
\alias{coro_goto}
\alias{coro_yield}
\alias{coro_return}
\title{Control flow for coroutine state machines}
\usage{
coro_goto(state, frame = caller_env())

coro_yield(state, value = NULL, frame = caller_env())

coro_return(value, frame = caller_env())
}
\description{
Control flow for coroutine state machines
}
\keyword{internal}
