% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iter.R
\name{done_box}
\alias{done_box}
\title{Box a final value to signal termination}
\usage{
done_box(x)
}
\arguments{
\item{x}{A final value that will be wrapped in a done box.}
}
\description{
A done box wraps a value to signal that a job is done. It is used
in flowery for two purposes:
}
\details{
\itemize{
\item From an \link[=iterator]{iterable function} to signal early
termination. If you know there is no more values in the iteration
you can return the last value in a "done box" to let the iterator
know that it is done. If you don't signal early, the iterator
might be reentered (at which point you can return \code{NULL} to
signal late termination).
}
}
