% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowery.R
\docType{package}
\name{flowery-package}
\alias{flowery}
\alias{flowery-package}
\title{flowery: Delimited Continuations in R}
\description{
This brings delimited continuations to R with a
computing-on-the-language approach.
}
\author{
\strong{Maintainer}: Lionel Henry \email{lionel@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
