% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{flowery_debug}
\alias{flowery_debug}
\title{Debug a generator or async function}
\usage{
flowery_debug(fn, value = TRUE)
}
\arguments{
\item{fn}{A generator factory or an async function created by
flowery.}

\item{value}{Whether to debug the function.}
}
\description{
\itemize{
\item Call \code{flowery_debug()} on a \code{\link[=generator]{generator()}}, \code{\link[=async]{async()}}, or
\code{\link[=async_generator]{async_generator()}} function to enable step-debugging.
\item Alternatively, set \code{options(flowery_debug = TRUE)} for
step-debugging through all functions created with flowery.
}
}
