% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterator.R
\name{iterate}
\alias{iterate}
\title{Iterate over an iterator}
\usage{
iterate(loop)
}
\arguments{
\item{loop}{A \code{for} loop expression.}
}
\description{
\code{iterate()} instruments \code{for} loops to support iteration with
flowery iterators and generators. See \code{vignette("generators")} for
more examples.
}
\examples{
iter <- gen(for (x in 1:10) yield(x^2))
iterate(for (x in iter) {
  print(x * 100)
})
}
\seealso{
\code{\link[=collect]{collect()}} is a functional that iterates over an
iterator and returns the values.
}
