% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterator.R
\name{iterator}
\alias{iterator}
\alias{exhausted}
\alias{is_exhausted}
\title{Iterator protocol}
\usage{
exhausted()

is_exhausted(x)
}
\arguments{
\item{x}{An object.}
}
\description{
An r-lib iterator is a function that implements the following protocol:
\itemize{
\item Calling the function advances the iterator. The new value is returned.
\item When the iterator is exhausted and there are no more elements to return, the symbol \code{quote(exhausted)} is returned. This signals exhaustion to the caller.
\item Once an iterator has signalled exhaustion, all subsequent invokations must consistently return \code{quote(exhausted)}.
}
}
