% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/async.R
\name{new_async}
\alias{new_async}
\title{Low-level constructor for async functions}
\usage{
new_async(fn, ops = NULL)
}
\description{
Unlike \code{\link[=async]{async()}} which uses concurrency based on
\href{https://rstudio.github.io/promises/}{promises}, \code{\link[=new_async]{new_async()}}
allows constructing async-await functions for other concurrency
frameworks.
}
\keyword{internal}
