% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/async.R
\name{new_async_generator}
\alias{new_async_generator}
\title{Low-level constructor for async functions}
\usage{
new_async_generator(fn, step, ops = NULL)
}
\arguments{
\item{step}{If \code{TRUE}, the async generator is immediately stepped
in after creation. This returns a promise. If \code{FALSE}, the async
generator is returned instead. Set \code{step} to \code{TRUE} when you
create an \code{\link[=async]{async()}} variant and to \code{FALSE} when you create an
\code{\link[=async_generator]{async_generator()}} variant.}
}
\description{
Unlike \code{\link[=async]{async()}} which uses concurrency based on
\href{https://rstudio.github.io/promises/}{promises},
\code{\link[=new_async_generator]{new_async_generator()}} allows constructing async-await functions
and async generators for other concurrency frameworks.
}
\keyword{internal}
