% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{new_gen_factory}
\alias{new_gen_factory}
\title{Create a generator factory}
\usage{
new_gen_factory(fn)
}
\arguments{
\item{fn}{A function that serves as mold for creating new
generators.}
}
\description{
Create a generator factory
}
