% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{new_generator}
\alias{new_generator}
\title{Create a generator}
\usage{
new_generator(body, env, args = list())
}
\arguments{
\item{body}{The body of the generator. It should use the `yield()`
keyword to pause the generator.}

\item{env}{The lexical scope of the generator.}

\item{args}{Variables passed to the generator.}
}
\description{
Create a generator
}
