% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step-reduce.R
\name{reduced_box}
\alias{reduced_box}
\alias{as_reduced_box}
\title{Create a boxed value for early termination of reduction}
\usage{
reduced_box(x)

as_reduced_box(x)
}
\arguments{
\item{x}{A value to box.}
}
\description{
A \link[rlang:box]{boxed} value of class \code{reduced_box} signals early
termination to \code{\link[=reduce_steps]{reduce_steps()}}. The boxed value is unboxed and
returned right away to the caller of \code{reduce_steps()}.
\itemize{
\item \code{reduced_box()} always boxes its input in a new box.
\item \code{as_reduced_box()} first checks if its input is a box of class
\code{reduced}. If it isn't, it boxes the input. Otherwise the input
is returned as is. This is useful to avoid double-boxing a value.
}
}
\examples{
box <- reduced_box(letters)

# Use `is_box(x, "reduced_box")` to check for a boxed value of type
# "reduced_box"
rlang::is_box(box, "reduced_box")
}
