% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step-reduce.R
\name{take}
\alias{take}
\alias{take_lgl}
\alias{take_int}
\alias{take_dbl}
\alias{take_cpl}
\alias{take_chr}
\alias{take_raw}
\title{Take n elements from a vector or iterator}
\usage{
take(.x, .n)

take_lgl(.x, .n)

take_int(.x, .n)

take_dbl(.x, .n)

take_cpl(.x, .n)

take_chr(.x, .n)

take_raw(.x, .n)
}
\arguments{
\item{.x}{A reducible object.}

\item{.n}{The number of elements to take from \code{.x}.}
}
\description{
These functions are similar to \code{head()} but also perform coercion
to a given output type. They support all reducible objects,
including iterators.
}
\examples{
take(letters, 5)
take_chr(1:10, 5)
}
\seealso{
\code{\link[=take_step]{take_step()}}
}
