% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iter-for.R
\name{as_iterator}
\alias{as_iterator}
\title{Transform vector to an iterator}
\usage{
as_iterator(x)
}
\arguments{
\item{x}{A vector to be transformed to an iterable function.
Functions are returned as is.}
}
\description{
Transform vector to an iterator
}
