% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step-reduce.R
\name{async_collect}
\alias{async_collect}
\title{Collect output of an asynchronous iterator}
\usage{
async_collect(x, n = NULL)
}
\arguments{
\item{x}{A reducible object.}

\item{n}{The number of elements to take from \code{x}.}
}
\description{
Collect output of an asynchronous iterator
}
