% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{package}
\name{coro-package}
\alias{coro}
\alias{coro-package}
\title{coro: Generators and transducers for R}
\description{
Make R a more coro language with generators and transducers.
}
\author{
\strong{Maintainer}: Lionel Henry \email{lionel@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
