% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine-loop.R, R/machine-switch.R
\name{coro_for_advance}
\alias{coro_for_advance}
\alias{coro_for_init}
\alias{coro_goto}
\alias{coro_yield}
\alias{coro_return}
\title{Control flow for coroutine state machines}
\usage{
coro_for_advance(elt_sym, iterator, env = caller_env())

coro_for_init(x)

coro_goto(state, frame = caller_env())

coro_yield(state, value = NULL, frame = caller_env())

coro_return(value, frame = caller_env())
}
\description{
Control flow for coroutine state machines
}
\keyword{internal}
