% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowery.R
\docType{package}
\name{flowery-package}
\alias{flowery}
\alias{flowery-package}
\title{flowery: Generators and transducers for R}
\description{
Make R a more flowery language with generators and transducers.
}
\author{
\strong{Maintainer}: Lionel Henry \email{lionel@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
