% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterator-for.R
\name{iterate}
\alias{iterate}
\title{Iterate over an iterator}
\usage{
iterate(loop)
}
\arguments{
\item{loop}{A \code{for} loop expression.}
}
\description{
\code{iterate()} instruments \code{for} loops to support iteration with
flowery iterators and generators.
}
\examples{
iter <- gen(for (x in 1:10) yield(x^2))
iterate(for (x in iter) {
  print(x * 100)
})
}
\seealso{
\code{\link[=iter_adapt]{iter_adapt()}} for a more functional way of dealing with
iterated valued.
}
