% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterator.R
\name{iterator}
\alias{iterator}
\alias{as_iterator}
\title{Iterable functions}
\usage{
as_iterator(x)
}
\arguments{
\item{x}{A vector to be transformed to an iterable function.
Functions are returned as is.}
}
\description{
A flowery iterator is a function that implements the iteration
protocol:
\itemize{
\item The iterator is advanced by invoking it without argument. This
returns the next value.
\item An iterator signals exhaustion by returning \code{NULL}.
}

The \code{NULL} sentinel synergises well with the R control flow
constructs like \verb{while ()} as they return \code{NULL} when they are
done.
}
